Refactor: Expression 역할 분리
